Refactor all game files into /game and its subfolders (#225)
* refactor IRtcHandler & VOIPDecorator into /voip

* refactor game base files into /game

* refactor all pre-game activities, fragments and adapters into /game/prep

* refactor handlers into /game/handler

* refactor timer into /game/handler

* refactor dialogs into /game/view
