+++ app/src/androidTest/java/com/github/fribourgsdp/radio/GameTest.kt
@@ -4,6 +4,7 @@ import com.github.fribourgsdp.radio.data.Genre
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.game.Game
 import org.junit.Test
 
 import org.junit.Assert.*
