+++ app/src/androidTest/java/com/github/fribourgsdp/radio/HostGameHandlerTest.kt
@@ -3,6 +3,8 @@ package com.github.fribourgsdp.radio
 import android.view.View
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.game.Game
+import com.github.fribourgsdp.radio.game.handler.HostGameHandler
 import com.github.fribourgsdp.radio.mockimplementations.FakeGameView
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.firestore.DocumentSnapshot
