+++ app/src/androidTest/java/com/github/fribourgsdp/radio/MainActivityTest.kt
@@ -11,6 +11,7 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.fribourgsdp.radio.config.SettingsActivity
 import com.github.fribourgsdp.radio.data.view.UserProfileActivity
+import com.github.fribourgsdp.radio.game.prep.GameSettingsActivity
 import junit.framework.Assert.assertEquals
 import org.hamcrest.Matchers
 import org.junit.After
