+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/LocalDatabase.kt
@@ -4,6 +4,7 @@ import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.game.Game
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.firestore.DocumentSnapshot
