+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockLobbyActivities.kt
@@ -6,6 +6,8 @@ import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.game.Game
+import com.github.fribourgsdp.radio.game.prep.*
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
