+++ app/src/main/java/com/github/fribourgsdp/radio/Database.kt
@@ -1,6 +1,7 @@
 package com.github.fribourgsdp.radio
 
 import com.github.fribourgsdp.radio.data.*
+import com.github.fribourgsdp.radio.game.Game
 import com.google.android.gms.tasks.Task
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.EventListener
