+++ app/src/main/java/com/github/fribourgsdp/radio/game/handler/HostGameHandler.kt
@@ -1,10 +1,12 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.game.handler
 
 import android.util.Log
+import com.github.fribourgsdp.radio.*
+import com.github.fribourgsdp.radio.game.Game
+import com.github.fribourgsdp.radio.game.GameView
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.FieldValue
 import java.lang.IllegalStateException
-import kotlin.streams.toList
 
 class HostGameHandler(private val game: Game, private val view: GameView, db: Database = FirestoreDatabase()): GameHandler(view, db) {
     private var latestSingerId: String? = null
