+++ app/src/main/java/com/github/fribourgsdp/radio/game/prep/LobbyActivity.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.game.prep
 
 import android.Manifest
 import android.content.Intent
@@ -11,10 +11,14 @@ import androidx.core.content.ContextCompat
 import androidx.fragment.app.DialogFragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.game.Game
+import com.github.fribourgsdp.radio.game.GameActivity
+import com.github.fribourgsdp.radio.game.view.QuitGameOrLobbyDialog
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
@@ -56,7 +60,7 @@ open class LobbyActivity : MyAppCompatActivity(){
     private lateinit var qrCodeDisplay: DialogFragment
 
     private var layoutManager:RecyclerView.LayoutManager? = null
-    private var layoutAdapter:RecyclerLobbyAdapter? = null
+    private var layoutAdapter: RecyclerLobbyAdapter? = null
 
     protected lateinit var mapIdToName: HashMap<String, String>
