+++ app/src/main/java/com/github/fribourgsdp/radio/game/view/LyricsPopup.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.game.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.fragment.app.DialogFragment
+import com.github.fribourgsdp.radio.R
 
 class LyricsPopup(private val lyrics : String) : DialogFragment() {
     private lateinit var closeButton : ImageView
