+++ app/src/main/java/com/github/fribourgsdp/radio/game/view/SongPickerDialog.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.game.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import androidx.fragment.app.DialogFragment
+import com.github.fribourgsdp.radio.R
+import com.github.fribourgsdp.radio.game.GameView
 
 
 class SongPickerDialog(private val choices: List<String>, private val listener: GameView.OnPickListener): DialogFragment() {
