+++ app/src/main/res/layout/activity_end_game.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EndGameActivity">
+    tools:context=".game.EndGameActivity">
 
     <TextView
         android:id="@+id/congratsTextView"
